requirements.txt
customtkinter
darkdetect
Pillow